+++ src/styles/Sections/contact.module.scss
@@ -1,4 +1,4 @@
-@import "config";
+@import "../config";
 
 .container {
   display: flex;
