+++ src/styles/Sections/skills.module.scss
@@ -1,4 +1,4 @@
-@import "config";
+@import "../config";
 
 .container {
   display: grid;
